parse down to minimal functionality
