+++ .changeset/sour-mails-cheer.md
@@ -2,4 +2,4 @@
 'ai': patch
 ---
 
-feat (ai/mcp): add support for `sessionId` and `strictMode` configuration to experimental MCP client
+chore (ai/mcp): add `assertCapability` method to experimental MCP client
