+++ content/docs/07-reference/01-ai-sdk-core/21-create-mcp-client.mdx
@@ -114,13 +114,6 @@ This feature is experimental and may change or be removed in the future.
               isOptional: true,
               description: 'Handler for uncaught errors',
             },
-            {
-              name: 'strictMode',
-              type: 'boolean',
-              isOptional: true,
-              description:
-                'Enforce strict capabilities mode. Defaults to false.',
-            },
           ],
         },
       ],
