+++ examples/mcp/package.json
@@ -7,7 +7,7 @@
     "sse:client": "tsx src/sse/client.ts",
     "stdio:build": "tsc src/stdio/server.ts --outDir src/stdio/dist --target es2023 --module nodenext",
     "stdio:client": "tsx src/stdio/client.ts",
-    "http:server": "tsx src/http/stateful-server.ts",
+    "http:server": "tsx src/http/server.ts",
     "http:client": "tsx src/http/client.ts",
     "custom-transport:build": "tsc src/custom-transport/server.ts --outDir src/custom-transport/dist --target es2023 --module nodenext",
     "custom-transport:client": "tsx src/custom-transport/client.ts",
