+++ packages/ai/core/tool/mcp/mock-mcp-transport.ts
@@ -28,7 +28,6 @@ export class MockMCPTransport implements MCPTransport {
   private failOnInvalidToolParams;
   private initializeResult;
   private sendError;
-  public sessionId;
 
   onmessage?: (message: JSONRPCMessage) => void;
   onclose?: () => void;
@@ -39,19 +38,16 @@ export class MockMCPTransport implements MCPTransport {
     failOnInvalidToolParams = false,
     initializeResult,
     sendError = false,
-    sessionId,
   }: {
     overrideTools?: MCPTool[];
     failOnInvalidToolParams?: boolean;
     initializeResult?: Record<string, unknown>;
     sendError?: boolean;
-    sessionId?: string;
   } = {}) {
     this.tools = overrideTools;
     this.failOnInvalidToolParams = failOnInvalidToolParams;
     this.initializeResult = initializeResult;
     this.sendError = sendError;
-    this.sessionId = sessionId;
   }
 
   async start(): Promise<void> {
