+++ packages/svelte/tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"version":"5.6.3"}
